Orleans 8 and refactor to match Orleans.Persistence.Redis
